lambda/discord_bot/requirements.txt
pynacl
boto3
